// project-management-frontend/src/@api/models/ai-elaboration-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project Management Tool API
 * API documentation for Project Management Tool
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AiElaborationResponseDto
 */
export interface AiElaborationResponseDto {
    /**
     * 
     * @type {string}
     * @memberof AiElaborationResponseDto
     */
    elaboratedTask?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof AiElaborationResponseDto
     */
    steps?: Array<string>;
}

/**
 * Check if a given object implements the AiElaborationResponseDto interface.
 */
export function instanceOfAiElaborationResponseDto(value: object): value is AiElaborationResponseDto {
    return true;
}

export function AiElaborationResponseDtoFromJSON(json: any): AiElaborationResponseDto {
    return AiElaborationResponseDtoFromJSONTyped(json, false);
}

export function AiElaborationResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): AiElaborationResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'elaboratedTask': json['elaboratedTask'] == null ? undefined : json['elaboratedTask'],
        'steps': json['steps'] == null ? undefined : json['steps'],
    };
}

export function AiElaborationResponseDtoToJSON(json: any): AiElaborationResponseDto {
    return AiElaborationResponseDtoToJSONTyped(json, false);
}

export function AiElaborationResponseDtoToJSONTyped(value?: AiElaborationResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'elaboratedTask': value['elaboratedTask'],
        'steps': value['steps'],
    };
}